Add snapshot module
